requirements.txt
################################################################################
# This requirements file has been automatically generated from `Pipfile` with
# `pipenv-to-requirements`
#
#
# This has been done to maintain backward compatibility with tools and services
# that do not support `Pipfile` yet.
#
# Do NOT edit it directly, use `pipenv install [-d]` to modify `Pipfile` and
# `Pipfile.lock` and then regenerate `requirements*.txt`.
################################################################################

blinker==1.8.2 ; python_version >= '3.8'
certifi==2024.7.4 ; python_version >= '3.6'
charset-normalizer==3.3.2 ; python_full_version >= '3.7.0'
click==8.1.7 ; python_version >= '3.7'
dash-bootstrap-components==1.6.0 ; python_version >= '3.8' and python_version < '4'
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash==2.17.1 ; python_version >= '3.8'
dnspython==2.6.1 ; python_version >= '3.8'
flask==3.0.3 ; python_version >= '3.8'
idna==3.7 ; python_version >= '3.5'
importlib-metadata==8.0.0 ; python_version >= '3.8'
itsdangerous==2.2.0 ; python_version >= '3.8'
jinja2==3.1.4 ; python_version >= '3.7'
markupsafe==2.1.5 ; python_version >= '3.7'
mysql-connector-python==9.0.0 ; python_version >= '3.8'
neo4j==5.22.0 ; python_version >= '3.7'
nest-asyncio==1.6.0 ; python_version >= '3.5'
numpy==2.0.1 ; python_version >= '3.12'
packaging==24.1 ; python_version >= '3.8'
pandas==2.2.2 ; python_version >= '3.9'
plotly==5.22.0 ; python_version >= '3.8'
pymongo==4.8.0 ; python_version >= '3.8'
python-dateutil==2.9.0.post0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1 ; python_version >= '3.8'
pytz==2024.1
requests==2.32.3 ; python_version >= '3.8'
retrying==1.3.4
setuptools==71.1.0 ; python_version >= '3.8'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tenacity==8.5.0 ; python_version >= '3.8'
typing-extensions==4.12.2 ; python_version >= '3.8'
tzdata==2024.1 ; python_version >= '2'
urllib3==2.2.2 ; python_version >= '3.8'
werkzeug==3.0.3 ; python_version >= '3.8'
zipp==3.19.2 ; python_version >= '3.8'
